ARTEMIS-1190 Long/int type mismatch in JDBCSequentialFile.setWritePosition

+++ artemis-jdbc-store/src/main/java/org/apache/activemq/artemis/jdbc/store/file/JDBCSequentialFile.java
@@ -77,7 +77,7 @@ public class JDBCSequentialFile implements SequentialFile {
       this.dbDriver = driver;
    }
 
-   void setWritePosition(int writePosition) {
+   void setWritePosition(long writePosition) {
       this.writePosition = writePosition;
    }
